+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.3.0 (2025-12-22)
+
+### Feat
+
+- biquad/peaking filter augmentations
+- NoiseAugmentation with per_example, per_batch, and fixed or range modes
+- auementations gain
+- auementations gain
+- gain augmentation with settings per batch/example/source/channel
+- handle additional dimensions
+- allow pedalboard to handle dims of 3,4
+
+### Fix
+
+- make sure pedalboard can't infinitify things
+- fix alls, imports, and nyquist
+- adds ruff, fixes lints, fixes bugs from lints
+- lpf and hpf augmentations were being applied to the whole batch! now per example
+- adds auementation store for gain
+- remove unnecessary code
+- fix groups in tests
+- update auementations store group for consistency
+
 ## 0.2.3 (2025-11-26)
 
 ### Fix